fix timeout overflow

+++ src/controller.h
@@ -83,7 +83,7 @@ void _toMode(Mode new_mode) {
 }
 } // namespace
 
-void hideAfter(uint16_t milliseconds) { timeout = millis() + milliseconds; }
+void hideAfter(uint32_t milliseconds) { timeout = millis() + milliseconds; }
 
 void turnOff() {
     Display::clear();
